refactor(indexer): extract dir-staleness check into isolated pass (R9)
Extract dir-staleness state logic from indexer.ts into a standalone,
testable pass to improve code clarity and enable focused integrity checks.
